moved database init script back to database.ts
